test({react,preact}-query/useQueries): move type-only tests to 'useQueries.test-d.tsx'
